Enable fullscreen SDL2 window
This patch allows the SDL2 window to be rendered in fullscreen.

While testing (on linux at least) I found that this does allow the
window to be fullscreen, but the frame buffer being drawn into the
window is incorrect. Only the top left corner of the frame is visible in
the fullscreen window.