Replace SDLKeyStrings with SDL_GetKeyName
SDLKeyStrings was a large hardcoded lookup table of key strings. The
values of the keycodes used as indices into the table are macros that
must have changed from SDL 1 to SDL2. This was causing a sigsegv when
trying to select "Configure" from the main menu. This patch replaces the
lookup table with the SDL function SDL_GetKeyName.